Move bignum, channel and validator modules to Domain module and re-export
