+++ sentry/src/domain.rs
@@ -1,0 +1,8 @@
+mod bignum;
+mod channel;
+mod validator;
+
+pub use bignum::BigNum;
+pub use channel::{Channel, ChannelSpec};
+pub use validator::ValidatorDesc;
+
